+++ hart/minions.py
@@ -84,6 +84,7 @@ def create_node(
         minion_config=None,
         connect_via_private_ip=False,
         no_public_ip=False,
+        non_interactive=False,
         **kwargs
         ):
     ssh_canary = utils.create_token()
@@ -112,7 +113,7 @@ def create_node(
         kwargs['zone'] = pick_distributed_zones(provider, region, roles)[0]
         print('Distributing minion to zone %s' % kwargs['zone'])
 
-    if not check_existing_minion(minion_id):
+    if not check_existing_minion(minion_id, non_interactive):
         print('Existing minions were found and did want to overwrite, aborting')
         return
 
@@ -224,7 +225,7 @@ def get_master_pubkey():
         return fh.read()
 
 
-def check_existing_minion(minion_id):
+def check_existing_minion(minion_id, non_interactive=False):
     minions = json.loads(subprocess.check_output([
         'salt-key',
         '--print', minion_id,
@@ -233,6 +234,10 @@ def check_existing_minion(minion_id):
 
     existing_categories = minions.keys()
     if existing_categories:
+        if non_interactive:
+            log_error('Existing minions matching %s were found in %s, not overwriting' % (
+                minion_id, ', '.join(existing_categories)))
+            return False
         should_continue = input('Existing minions matching %s were found in %s, '
             'overwrite? [y/N]' % (minion_id, ', '.join(existing_categories)))
         return should_continue == 'y'